PP-15510 move fixture to model package
