+++ src/test/java/uk/gov/pay/connector/gatewayaccount/model/AdyenAccountSetupTaskEntityFixture.java
@@ -1,9 +1,5 @@
-package uk.gov.pay.connector.gatewayaccount.service;
+package uk.gov.pay.connector.gatewayaccount.model;
 
-import uk.gov.pay.connector.gatewayaccount.model.AdyenAccountSetupStatus;
-import uk.gov.pay.connector.gatewayaccount.model.AdyenAccountSetupTask;
-import uk.gov.pay.connector.gatewayaccount.model.AdyenAccountSetupTaskEntity;
-import uk.gov.pay.connector.gatewayaccount.model.GatewayAccountEntity;
 import uk.gov.pay.connector.gatewayaccountcredentials.model.GatewayAccountCredentialsEntity;
 
 import java.util.ArrayList;
@@ -16,11 +12,11 @@ public class AdyenAccountSetupTaskEntityFixture {
     private AdyenAccountSetupTask task;
     private AdyenAccountSetupStatus status;
     
-    static AdyenAccountSetupTaskEntityFixture anAdyenAccountSetupTaskEntityFixture() {
+    public static AdyenAccountSetupTaskEntityFixture anAdyenAccountSetupTaskEntityFixture() {
         return new AdyenAccountSetupTaskEntityFixture();
     }
 
-    static ArrayList<AdyenAccountSetupTaskEntity> anAdyenAccountSetupTaskEntityListWithAllTasksCompleted(GatewayAccountEntity gatewayAccount, GatewayAccountCredentialsEntity gatewayAccountCredential) {
+    public static ArrayList<AdyenAccountSetupTaskEntity> anAdyenAccountSetupTaskEntityListWithAllTasksCompleted(GatewayAccountEntity gatewayAccount, GatewayAccountCredentialsEntity gatewayAccountCredential) {
         var adyenAccountSetupTaskEntities = new ArrayList<AdyenAccountSetupTaskEntity>();
         
         Arrays.stream(AdyenAccountSetupTask.values()).forEach(adyenAccountSetupTask -> adyenAccountSetupTaskEntities
